+++ src/main/java/com/secuconnect/client/ApiException.java
@@ -1,4 +1,3 @@
-
 package com.secuconnect.client;
 
 import com.google.gson.Gson;